rename: 트레이닝 관련 controller,service 파일 이름 변경 및 분리
