+++ src/main/java/com/fithub/fithubbackend/domain/Training/api/TrainingController.java
@@ -1,19 +1,17 @@
 package com.fithub.fithubbackend.domain.Training.api;
 
 import com.fithub.fithubbackend.domain.Training.application.TrainingService;
-import com.fithub.fithubbackend.domain.Training.dto.TrainersReserveInfoDto;
-import com.fithub.fithubbackend.domain.Training.dto.TrainingCreateDto;
-import com.fithub.fithubbackend.domain.user.domain.User;
-import com.fithub.fithubbackend.global.domain.AuthUser;
-import com.fithub.fithubbackend.global.exception.CustomException;
-import com.fithub.fithubbackend.global.exception.ErrorCode;
+import com.fithub.fithubbackend.domain.Training.dto.TrainingInfoDto;
+import com.fithub.fithubbackend.domain.Training.dto.TrainingOutlineDto;
+import com.fithub.fithubbackend.domain.Training.dto.TrainingReviewDto;
+import com.fithub.fithubbackend.domain.Training.dto.TrainingSearchConditionDto;
 import com.fithub.fithubbackend.global.exception.ErrorResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import jakarta.validation.Valid;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -22,111 +20,52 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
+@Tag(name = "training (트레이닝)", description = "인증이 필요없는 트레이닝 api")
 @RestController
-@RequestMapping("/trainer/training")
+@RequestMapping("/training")
 @RequiredArgsConstructor
 public class TrainingController {
-
     private final TrainingService trainingService;
 
-    @Operation(summary = "트레이닝 생성, swagger에서 테스트 불가능, 이미지는 모두 images로 주면 됨", responses = {
-            @ApiResponse(responseCode = "200", description = "생성됨"),
-            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "409", description = "예약 가능 날짜에 현재보다 이전 날짜가 들어있음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-    })
-    @PostMapping
-    public ResponseEntity<Long> createTraining(@Valid TrainingCreateDto dto, @AuthUser User user) {
-        if(user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-        return ResponseEntity.ok(trainingService.createTraining(dto, user));
+    @Operation(summary = "트레이닝 전체 조회, page 사용 (size = 9, sort=\"id\" desc 적용되어있음. swagger에서 보낼 때 따로 지정하는 거 없으면 parameter에 pageable은 다 지우고 보내도 됨)")
+    @GetMapping("/all")
+    public ResponseEntity<Page<TrainingOutlineDto>> searchAll(@PageableDefault(size = 9, sort = "id",  direction = Sort.Direction.DESC) Pageable pageable) {
+        return ResponseEntity.ok(trainingService.searchAll(pageable));
     }
 
-//    @Operation(summary = "트레이닝 수정", responses = {
-//            @ApiResponse(responseCode = "200", description = "수정됨"),
-//            @ApiResponse(responseCode = "400", description = "수정 날짜에 현재보다 이전 날짜가 들어있음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class)))
-//            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-//            @ApiResponse(responseCode = "409", description = "마감 처리 된 트레이닝을 수정하려고 함", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-//    })
-//    @PutMapping
-//    public ResponseEntity<Long> updateTraining(@RequestBody @Valid TrainingCreateDto dto, @RequestParam Long trainingId, @AuthUser User user) {
-//        // TODO: 예약 관련 기능 끝난 후 수정 필요
-//        if (user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-//        return ResponseEntity.ok(trainingService.updateTraining(dto, trainingId, user.getEmail()));
-//    }
-
-    @Operation(summary = "트레이닝 삭제 (soft delete)", parameters = {
-            @Parameter(name = "trainingId", description = "삭제할 트레이닝 id")
-    }, responses = {
-            @ApiResponse(responseCode = "200", description = "삭제 성공"),
-            @ApiResponse(responseCode = "400", description = "해당 트레이닝에 취소/완료되지 않은 예약이 남아 삭제 불가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "403", description = "해당 트레이닝을 작성한 트레이너가 아니라 삭제 권한이 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "id에 해당하는 트레이닝이 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-    })
-    @DeleteMapping
-    public ResponseEntity<String> deleteTraining(@RequestParam Long trainingId, @AuthUser User user) {
-        if(user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-        trainingService.deleteTraining(trainingId, user.getEmail());
-        return ResponseEntity.ok().body("완료");
-    }
-
-    @Operation(summary = "트레이닝 수동 마감 설정", parameters = {
-            @Parameter(name = "trainingId", description = "마감할 트레이닝 id")
+    @Operation(summary = "트레이닝 하나 상세 조회", parameters = {
+            @Parameter(name = "trainingId", description = "조회할 트레이닝의 primary key(id)")
     }, responses = {
             @ApiResponse(responseCode = "200", description = "조회 성공"),
-            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "403", description = "해당 트레이닝을 작성한 트레이너가 아니라 수정 권한이 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "id에 해당하는 트레이닝이 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
+            @ApiResponse(responseCode = "400", description = "삭제된 트레이닝", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
+            @ApiResponse(responseCode = "404", description = "해당하는 트레이닝 존재 X", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class)))
     })
-    @PutMapping("/close")
-    public ResponseEntity<String> closeTraining(@RequestParam Long trainingId, @AuthUser User user) {
-        if(user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-        trainingService.closeTraining(trainingId, user);
-        return ResponseEntity.ok().body("완료");
+    @GetMapping
+    public ResponseEntity<TrainingInfoDto> searchById(@RequestParam Long trainingId) {
+        return ResponseEntity.ok(trainingService.searchById(trainingId));
     }
 
-    @Operation(summary = "트레이닝 수동 마감 해제", parameters = {
-            @Parameter(name = "trainingId", description = "오픈할 트레이닝 id")
+    @Operation(summary = "트레이닝 상세 조회에서 트레이닝 리뷰 리스트 조회", parameters = {
+            @Parameter(name = "trainingId", description = "조회할 트레이닝의 primary key(id)")
     }, responses = {
             @ApiResponse(responseCode = "200", description = "조회 성공"),
-            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "403", description = "해당 트레이닝을 작성한 트레이너가 아니라 수정 권한이 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "id에 해당하는 트레이닝이 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "409", description = "트레이닝 예약 가능한 마지막 날짜가 현재보다 이전이므로 수정 불가능. 트레이닝 수정을 통해 날짜 추가 필요", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
     })
-    @PutMapping("/open")
-    public ResponseEntity<String> openTraining(@RequestParam Long trainingId, @AuthUser User user) {
-        if(user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-        trainingService.openTraining(trainingId, user);
-        return ResponseEntity.ok().body("완료");
+    @GetMapping("/reviews")
+    public ResponseEntity<List<TrainingReviewDto>> getTrainingReviews (@RequestParam Long trainingId) {
+        return ResponseEntity.ok(trainingService.getTrainingReviews(trainingId));
     }
 
-    @Operation(summary = "트레이너의 예약 리스트 조회", parameters = {
-            @Parameter(name = "pageable", description = "조회할 목록의 page, size, sort(기본은 id (생성 순), 예약된 트레이닝 날짜 순은 reserveDateTime으로 주면 됨)")
+    @Operation(summary = "트레이닝 검색 (필터 포함)", parameters = {
+            @Parameter(name = "conditions", description = "검색 조건들 (제목 키워드, 최저/최고가, 시작/마감일), 사용할 조건만 보내야됨"),
+            @Parameter(name = "pageable", description = "조회할 목록의 page, size, sort(기본은 id desc(최신 생성 순) 변경 안 할거면 sort 부분은 지우기, title, startDate, endDate, price 지정 가능)")
     }, responses = {
             @ApiResponse(responseCode = "200", description = "조회 성공"),
-            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "해당 회원은 트레이너가 아님", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
     })
-    @GetMapping("/reservations")
-    public ResponseEntity<Page<TrainersReserveInfoDto>> getReservationList(@AuthUser User user,
-                                                                           @PageableDefault(sort="id", direction = Sort.Direction.DESC) Pageable pageable) {
-        if(user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-        return ResponseEntity.ok(trainingService.getReservationList(user, pageable));
-    }
-
-    @Operation(summary = "트레이너의 예약 노쇼 처리", parameters = {
-            @Parameter(name = "reservationId", description = "상태를 변경하려는 예약의 id")
-    }, responses = {
-            @ApiResponse(responseCode = "200", description = "노쇼 처리 성공"),
-            @ApiResponse(responseCode = "400", description = "해당 트레이닝이 완료 상태가 아님 (노쇼는 완료 상태인 예약만 가능, 완료 상태는 정각마다 스케줄러에 의해 변경됨)", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "401", description = "로그인한 사용자만 가능", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "403", description = "해당 트레이닝을 생성한 트레이너가 아님. 권한 없음", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "존재하지 않는 예약", content = @Content(schema = @Schema(implementation = ErrorResponseDto.class))),
-    })
-    @PutMapping("/reservation/status/noshow")
-    public ResponseEntity<String> updateReservationStatusNoShow(@AuthUser User user, @RequestParam Long reservationId) {
-        if (user == null) throw new CustomException(ErrorCode.AUTHENTICATION_ERROR, "로그인한 사용자만 가능합니다.");
-        trainingService.updateReservationStatusNoShow(user.getEmail(), reservationId);
-        return ResponseEntity.ok().body("완료");
+    @PostMapping("/search")
+    public ResponseEntity<Page<TrainingOutlineDto>> searchTrainingByConditions(@RequestBody TrainingSearchConditionDto conditions
+            , @PageableDefault(size = 9, sort = "id",  direction = Sort.Direction.DESC) Pageable pageable) {
+        return ResponseEntity.ok(trainingService.searchTrainingByConditions(conditions, pageable));
     }
 }
