+++ src/main/java/com/fithub/fithubbackend/domain/Training/application/UserTrainingReservationService.java
@@ -7,18 +7,10 @@ import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
-public interface UserTrainingService {
-    Page<TrainingOutlineDto> searchAll(Pageable pageable);
-    TrainingInfoDto searchById(Long id);
-    List<TrainingReviewDto> getTrainingReviews(Long id);
+public interface UserTrainingReservationService {
 
-    boolean isLikesTraining(Long trainingId, User user);
-    void likesTraining(Long trainingId, User user);
-    void cancelTrainingLikes(Long trainingId, User user);
 
-    List<TrainingLikesInfoDto> getTrainingLikesList(User user);
 
-    Page<TrainingOutlineDto> searchTrainingByConditions(TrainingSearchConditionDto conditions, Pageable pageable);
 
     Page<UsersReserveInfoDto> getTrainingReservationList(User user, Pageable pageable);
