+++ src/main/java/com/fithub/fithubbackend/domain/Training/application/UserTrainingReservationServiceImpl.java
@@ -1,12 +1,13 @@
 package com.fithub.fithubbackend.domain.Training.application;
 
 import com.fithub.fithubbackend.domain.Training.domain.ReserveInfo;
-import com.fithub.fithubbackend.domain.Training.domain.Training;
-import com.fithub.fithubbackend.domain.Training.domain.TrainingLikes;
 import com.fithub.fithubbackend.domain.Training.domain.TrainingReview;
-import com.fithub.fithubbackend.domain.Training.dto.*;
+import com.fithub.fithubbackend.domain.Training.dto.TrainingReviewReqDto;
+import com.fithub.fithubbackend.domain.Training.dto.UsersReserveInfoDto;
+import com.fithub.fithubbackend.domain.Training.dto.UsersTrainingReviewDto;
 import com.fithub.fithubbackend.domain.Training.enums.ReserveStatus;
-import com.fithub.fithubbackend.domain.Training.repository.*;
+import com.fithub.fithubbackend.domain.Training.repository.ReserveInfoRepository;
+import com.fithub.fithubbackend.domain.Training.repository.TrainingReviewRepository;
 import com.fithub.fithubbackend.domain.user.domain.User;
 import com.fithub.fithubbackend.global.exception.CustomException;
 import com.fithub.fithubbackend.global.exception.ErrorCode;
@@ -20,97 +21,10 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class UserTrainingServiceImpl implements UserTrainingService {
+public class UserTrainingReservationServiceImpl implements UserTrainingReservationService {
 
-    private final TrainingRepository trainingRepository;
-    private final TrainingLikesRepository trainingLikesRepository;
     private final ReserveInfoRepository reserveInfoRepository;
     private final TrainingReviewRepository trainingReviewRepository;
-
-    private final CustomTrainingRepository customTrainingRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public Page<TrainingOutlineDto> searchAll(Pageable pageable) {
-        Page<Training> trainingPage = trainingRepository.findAllByDeletedFalse(pageable);
-        return trainingPage.map(TrainingOutlineDto::toDto);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public TrainingInfoDto searchById(Long id) {
-        Training training = trainingRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND, "해당하는 트레이닝이 존재하지 않습니다."));
-        if (training.isDeleted()) {
-            throw new CustomException(ErrorCode.BAD_REQUEST, "삭제된 트레이닝입니다.");
-        }
-
-        TrainingInfoDto dto = TrainingInfoDto.toDto(training);
-        if (training.getImages() != null && !training.getImages().isEmpty()) {
-            List<TrainingDocumentDto> images = training.getImages().stream().map(TrainingDocumentDto::toDto).toList();
-            dto.updateImages(images);
-        }
-        return dto;
-    }
-
-    @Override
-    public List<TrainingReviewDto> getTrainingReviews(Long id) {
-        List<TrainingReview> trainingReviewList = trainingReviewRepository.findByTrainingId(id);
-        return trainingReviewList.stream().map(TrainingReviewDto::toDto).toList();
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public boolean isLikesTraining(Long trainingId, User user) {
-        if (!trainingRepository.existsById(trainingId)) {
-            throw new CustomException(ErrorCode.NOT_FOUND, "존재하지 않는 트레이닝입니다.");
-        }
-        return trainingLikesRepository.existsByTrainingIdAndUserId(trainingId, user.getId());
-    }
-
-    @Override
-    @Transactional
-    public void likesTraining(Long trainingId, User user) {
-        Training training = trainingRepository.findById(trainingId).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND, "존재하지 않는 트레이닝입니다."));
-        checkClosed(training.isClosed());
-
-        if (user.getId().equals(training.getTrainer().getUser().getId())) {
-            throw new CustomException(ErrorCode.UNKNOWN_ERROR, "트레이너는 자신의 트레이닝을 찜할 수 없습니다.");
-        }
-
-        TrainingLikes trainingLikes = TrainingLikes.builder().training(training).user(user).build();
-        trainingLikesRepository.save(trainingLikes);
-    }
-
-    private void checkClosed(boolean closed) {
-        if (closed) {
-            throw new CustomException(ErrorCode.UNCORRECTABLE_DATA, "마감된 트레이닝은 찜할 수 없습니다.");
-        }
-    }
-
-    @Override
-    @Transactional
-    public void cancelTrainingLikes(Long trainingId, User user) {
-        if (!trainingRepository.existsById(trainingId)) {
-            throw new CustomException(ErrorCode.NOT_FOUND, "존재하지 않는 트레이닝입니다.");
-        }
-        TrainingLikes trainingLikes = trainingLikesRepository.findByTrainingIdAndUserId(trainingId, user.getId()).orElseThrow(() -> new CustomException(ErrorCode.UNCORRECTABLE_DATA, "트레이닝을 찜하지 않았습니다."));
-        trainingLikesRepository.delete(trainingLikes);
-    }
-
-    @Override
-    public List<TrainingLikesInfoDto> getTrainingLikesList(User user) {
-        List<TrainingLikes> trainingLikes = trainingLikesRepository.findByUserId(user.getId());
-        return trainingLikes.stream().map(t -> TrainingLikesInfoDto.builder()
-                .id(t.getId())
-                .trainingOutlineDto(TrainingOutlineDto.toDto(t.getTraining()))
-                .build()).toList();
-    }
-
-    @Override
-    public Page<TrainingOutlineDto> searchTrainingByConditions(TrainingSearchConditionDto conditions, Pageable pageable) {
-        Page<Training> trainingList = customTrainingRepository.searchByConditions(conditions, pageable);
-        return trainingList.map(TrainingOutlineDto::toDto);
-    }
 
     @Override
     public Page<UsersReserveInfoDto> getTrainingReservationList(User user, Pageable pageable) {
